+++ infopark_reactor/lib/reactor/cm/editorial_group.rb
@@ -15,6 +15,10 @@ module Reactor
 
       # Overwritten method from +Group+.
       def base_name
+        self.class.base_name
+      end
+
+      def self.base_name
         'groupProxy'
       end